refactor(hydro_lang)!: rename `graph_util` => `graph::config`
Clearer name, and all the declarations in it were for configuring graph generation anyways.
