+++ hydro_test/examples/benchmark_paxos.rs
@@ -18,7 +18,7 @@ async fn main() {
     #[command(author, version, about, long_about = None)]
     struct BenchmarkArgs {
         #[command(flatten)]
-        graph: hydro_lang::graph_util::GraphConfig,
+        graph: hydro_lang::graph::config::GraphConfig,
 
         /// Use GCP for deployment (provide project name)
         #[arg(long)]
