+++ hydro_test/examples/chat.rs
@@ -2,7 +2,7 @@ use clap::Parser;
 use dfir_rs::tokio_util::codec::LinesCodec;
 use hydro_deploy::Deployment;
 use hydro_lang::deploy::TrybuildHost;
-use hydro_lang::graph_util::GraphConfig;
+use hydro_lang::graph::config::GraphConfig;
 use hydro_lang::{Location, NetworkHint, nondet};
 
 #[derive(Parser, Debug)]
